// workflow_fe/src/lib/api/gen/types/ResubmitLeaveDto.ts
/**
 * Generated by Kubb (https://kubb.dev/).
 * Do not edit manually.
 */

export type ResubmitLeaveDto = {
  /**
   * @type string | undefined
   */
  leaveType?: string;
  /**
   * @minLength 1
   * @type number | undefined
   */
  leaveDays?: number;
  /**
   * @type string | undefined
   */
  startDate?: string;
  /**
   * @type string | undefined
   */
  endDate?: string;
  /**
   * @type string | undefined
   */
  reason?: string;
  /**
   * @type string | undefined
   */
  employeeGrade?: string;
  /**
   * @type string | undefined
   */
  departmentId?: string;
  /**
   * @type object | undefined
   */
  customFieldsJson?: object;
};